Reshuffle files
