+++ src/eko/evolution_operator/__init__.py
@@ -17,12 +17,6 @@ from scipy import LowLevelCallable, integrate
 import ekore.anomalous_dimensions.polarized.space_like as ad_ps
 import ekore.anomalous_dimensions.unpolarized.space_like as ad_us
 import ekore.anomalous_dimensions.unpolarized.time_like as ad_ut
-from ekore.cffilib import (
-    c_quad_ker_qcd,
-    im_double_address,
-    re_double_address,
-    read_complex,
-)
 
 from .. import basis_rotation as br
 from .. import interpolation, mellin
@@ -35,91 +29,9 @@ from ..kernels import utils
 from ..kernels import valence_qed as qed_v
 from ..matchings import Segment
 from ..member import OpMember
+from .quad_ker import QuadCargs, c_quad_ker_qcd, cb_quad_ker_qcd, cb_quad_ker_qcd_T
 
 logger = logging.getLogger(__name__)
-
-
-@nb.njit(cache=True)
-def select_singlet_element(ker, mode0, mode1):
-    """Select element of the singlet matrix.
-
-    Parameters
-    ----------
-    ker : numpy.ndarray
-        singlet integration kernel
-    mode0 : int
-        id for first sector element
-    mode1 : int
-        id for second sector element
-
-    Returns
-    -------
-    complex
-        singlet integration kernel element
-    """
-    k = 0 if mode0 == 100 else 1
-    l = 0 if mode1 == 100 else 1
-    return ker[k, l]
-
-
-@nb.njit(cache=True)
-def select_QEDsinglet_element(ker, mode0, mode1):
-    """Select element of the QEDsinglet matrix.
-
-    Parameters
-    ----------
-    ker : numpy.ndarray
-        QEDsinglet integration kernel
-    mode0 : int
-        id for first sector element
-    mode1 : int
-        id for second sector element
-    Returns
-    -------
-    ker : complex
-        QEDsinglet integration kernel element
-    """
-    if mode0 == 21:
-        index1 = 0
-    elif mode0 == 22:
-        index1 = 1
-    elif mode0 == 100:
-        index1 = 2
-    else:
-        index1 = 3
-    if mode1 == 21:
-        index2 = 0
-    elif mode1 == 22:
-        index2 = 1
-    elif mode1 == 100:
-        index2 = 2
-    else:
-        index2 = 3
-    return ker[index1, index2]
-
-
-@nb.njit(cache=True)
-def select_QEDvalence_element(ker, mode0, mode1):
-    """
-    Select element of the QEDvalence matrix.
-
-    Parameters
-    ----------
-    ker : numpy.ndarray
-        QEDvalence integration kernel
-    mode0 : int
-        id for first sector element
-    mode1 : int
-        id for second sector element
-    Returns
-    -------
-    ker : complex
-        QEDvalence integration kernel element
-    """
-    index1 = 0 if mode0 == 10200 else 1
-    index2 = 0 if mode1 == 10200 else 1
-    return ker[index1, index2]
-
 
 spec = [
     ("is_singlet", nb.boolean),
@@ -190,611 +102,6 @@ class QuadKerBase:
         if pj == 0.0:
             return 0.0
         return self.path.prefactor * pj * self.path.jac
-
-
-@nb.njit(cache=False)
-def quad_ker(
-    u,
-    order,
-    mode0,
-    mode1,
-    method,
-    is_log,
-    logx,
-    areas,
-    as_list,
-    mu2_from,
-    mu2_to,
-    a_half,
-    alphaem_running,
-    nf,
-    L,
-    ev_op_iterations,
-    ev_op_max_order,
-    sv_mode,
-    is_threshold,
-    n3lo_ad_variation,
-    is_polarized,
-    is_time_like,
-    quad_ker_qcd_ns_address,
-):
-    """Raw evolution kernel inside quad.
-
-    Parameters
-    ----------
-    u : float
-        quad argument
-    order : int
-        perturbation order
-    mode0: int
-        pid for first sector element
-    mode1 : int
-        pid for second sector element
-    method : str
-        method
-    is_log : boolean
-        is a logarithmic interpolation
-    logx : float
-        Mellin inversion point
-    areas : tuple
-        basis function configuration
-    as1 : float
-        target coupling value
-    as0 : float
-        initial coupling value
-    mu2_from : float
-        initial value of mu2
-    mu2_from : float
-        final value of mu2
-    aem_list : list
-        list of electromagnetic coupling values
-    alphaem_running : bool
-        whether alphaem is running or not
-    nf : int
-        number of active flavors
-    L : float
-        logarithm of the squared ratio of factorization and renormalization scale
-    ev_op_iterations : int
-        number of evolution steps
-    ev_op_max_order : int
-        perturbative expansion order of U
-    sv_mode: int, `enum.IntEnum`
-        scale variation mode, see `eko.scale_variations.Modes`
-    is_threshold : boolean
-        is this an intermediate threshold operator?
-    n3lo_ad_variation : tuple
-        |N3LO| anomalous dimension variation ``(gg_var, gq_var, qg_var, qq_var)``
-    is_polarized : boolean
-        is polarized evolution ?
-    is_time_like : boolean
-        is time-like evolution ?
-
-    Returns
-    -------
-    float
-        evaluated integration kernel
-    """
-    ker_base = QuadKerBase(u, is_log, logx, mode0)
-    integrand = ker_base.integrand(areas)
-    if integrand == 0.0:
-        return 0.0
-    if order[1] == 0:
-        ker = quad_ker_qcd(
-            ker_base,
-            order,
-            mode0,
-            mode1,
-            method,
-            as_list[-1],
-            as_list[0],
-            nf,
-            L,
-            ev_op_iterations,
-            ev_op_max_order,
-            sv_mode,
-            is_threshold,
-            is_polarized,
-            is_time_like,
-            n3lo_ad_variation,
-            quad_ker_qcd_ns_address,
-        )
-    else:
-        ker = quad_ker_qed(
-            ker_base,
-            order,
-            mode0,
-            mode1,
-            method,
-            as_list,
-            mu2_from,
-            mu2_to,
-            a_half,
-            alphaem_running,
-            nf,
-            L,
-            ev_op_iterations,
-            ev_op_max_order,
-            sv_mode,
-            is_threshold,
-        )
-
-    # recombine everything
-    return np.real(ker * integrand)
-
-
-@nb.njit(cache=False)
-def quad_ker_qcd(
-    ker_base,
-    order,
-    mode0,
-    mode1,
-    method,
-    as1,
-    as0,
-    nf,
-    L,
-    ev_op_iterations,
-    ev_op_max_order,
-    sv_mode,
-    is_threshold,
-    is_polarized,
-    is_time_like,
-    n3lo_ad_variation,
-    quad_ker_qcd_ns_address,
-):
-    """Raw evolution kernel inside quad.
-
-    Parameters
-    ----------
-    quad_ker : float
-        quad argument
-    order : int
-        perturbation order
-    mode0: int
-        pid for first sector element
-    mode1 : int
-        pid for second sector element
-    method : str
-        method
-    as1 : float
-        target coupling value
-    as0 : float
-        initial coupling value
-    nf : int
-        number of active flavors
-    L : float
-        logarithm of the squared ratio of factorization and renormalization scale
-    ev_op_iterations : int
-        number of evolution steps
-    ev_op_max_order : int
-        perturbative expansion order of U
-    sv_mode: int, `enum.IntEnum`
-        scale variation mode, see `eko.scale_variations.Modes`
-    is_threshold : boolean
-        is this an itermediate threshold operator?
-    n3lo_ad_variation : tuple
-        |N3LO| anomalous dimension variation ``(gg_var, gq_var, qg_var, qq_var)``
-
-    Returns
-    -------
-    float
-        evaluated integration kernel
-    """
-    # compute the actual evolution kernel for pure QCD
-    if ker_base.is_singlet:
-        if is_polarized:
-            if is_time_like:
-                raise NotImplementedError("Polarized, time-like is not implemented")
-            else:
-                gamma_singlet = ad_ps.gamma_singlet(order, ker_base.n, nf)
-        else:
-            if is_time_like:
-                gamma_singlet = ad_ut.gamma_singlet(order, ker_base.n, nf)
-            else:
-                gamma_singlet = ad_us.c_gamma_singlet(
-                    order, ker_base.n, nf, n3lo_ad_variation
-                )
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_singlet = sv.exponentiated.gamma_variation(
-                gamma_singlet, order, nf, L
-            )
-        ker = s.dispatcher(
-            order,
-            method,
-            gamma_singlet,
-            as1,
-            as0,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv.expanded.singlet_variation(gamma_singlet, as1, order, nf, L, dim=2)
-            ) @ np.ascontiguousarray(ker)
-        ker = select_singlet_element(ker, mode0, mode1)
-    else:
-        # if is_polarized:
-        #     if is_time_like:
-        #         raise NotImplementedError("Polarized, time-like is not implemented")
-        #     else:
-        #         gamma_ns = ad_ps.gamma_ns(order, mode0, ker_base.n, nf)
-        # else:
-        #     if is_time_like:
-        #         gamma_ns = ad_ut.gamma_ns(order, mode0, ker_base.n, nf)
-        #     else:
-        #         gamma_ns = ad_us.c_gamma_ns(order, mode0, ker_base.n, nf)
-        ker = quad_ker_qcd_ns(
-            order,
-            mode0,
-            ker_base.n,
-            nf,
-            quad_ker_qcd_ns_address,
-            L,
-            as1,
-            as0,
-            ev_op_iterations,
-            is_threshold,
-        )
-    return ker
-
-
-cb_quad_ker_qcd_T = ctypes.CFUNCTYPE(
-    ctypes.c_double,
-    ctypes.POINTER(ctypes.c_double),  # re_gamma_ns_raw
-    ctypes.POINTER(ctypes.c_double),  # im_gamma_ns_raw
-    ctypes.c_double,  # re_n
-    ctypes.c_double,  # im_n
-    ctypes.c_double,  # re_jac
-    ctypes.c_double,  # im_jac
-    ctypes.c_uint,  # order_qcd
-    ctypes.c_bool,  # is_singlet
-    ctypes.c_uint,  # mode0
-    ctypes.c_uint,  # mode1
-    ctypes.c_uint,  # nf
-    ctypes.c_bool,  # is_log
-    ctypes.c_double,  # logx
-    ctypes.c_double * 12,  # areas_raw
-    ctypes.c_uint,  # polynomial_degreee
-    ctypes.c_double,  # L
-    ctypes.c_uint,  # method_num
-    ctypes.c_double,  # as1
-    ctypes.c_double,  # as0
-    ctypes.c_uint,  # ev_op_iterations
-    ctypes.c_uint,  # ev_op_max_order_qcd
-    ctypes.c_uint,  # sv_mode_num
-    ctypes.c_bool,  # is_threshold
-)
-
-libc = ctypes.CDLL("./ekorepp.so")
-c_quad_ker_qcd2 = libc.c_quad_ker_qcd
-
-
-class QuadCargs(ctypes.Structure):
-    """Arguments to C call."""
-
-    _fields_ = [
-        ("order_qcd", ctypes.c_uint),
-        ("mode0", ctypes.c_uint),
-        ("mode1", ctypes.c_uint),
-        ("is_polarized", ctypes.c_bool),
-        ("is_time_like", ctypes.c_bool),
-        ("nf", ctypes.c_uint),
-        ("py", cb_quad_ker_qcd_T),
-        ("is_log", ctypes.c_bool),
-        ("logx", ctypes.c_double),
-        ("areas", ctypes.POINTER(ctypes.c_double)),
-        ("areas_x", ctypes.c_uint),
-        ("areas_y", ctypes.c_uint),
-        ("L", ctypes.c_double),
-        ("method_num", ctypes.c_uint),
-        ("as1", ctypes.c_double),
-        ("as0", ctypes.c_double),
-        ("ev_op_iterations", ctypes.c_uint),
-        ("ev_op_max_order_qcd", ctypes.c_uint),
-        ("sv_mode_num", ctypes.c_uint),
-        ("is_threshold", ctypes.c_bool),
-    ]
-
-
-c_quad_ker_qcd2.argtypes = [ctypes.c_double, ctypes.c_void_p]
-c_quad_ker_qcd2.restype = ctypes.c_double
-
-
-@nb.cfunc(
-    nb.types.double(
-        nb.types.CPointer(nb.types.double),  # re_gamma_ns_raw
-        nb.types.CPointer(nb.types.double),  # im_gamma_ns_raw
-        nb.types.double,  # re_n
-        nb.types.double,  # im_n
-        nb.types.double,  # re_jac
-        nb.types.double,  # im_jac
-        nb.types.uint,  # order_qcd
-        nb.types.bool_,  # is_singlet
-        nb.types.uint,  # mode0
-        nb.types.uint,  # mode1
-        nb.types.uint,  # nf
-        nb.types.bool_,  # is_log
-        nb.types.double,  # logx
-        nb.types.CPointer(nb.types.double),  # areas_raw
-        nb.types.uint,  # areas_x
-        nb.types.uint,  # areas_y
-        nb.types.double,  # L
-        nb.types.uint,  # method_num
-        nb.types.double,  # as1
-        nb.types.double,  # as0
-        nb.types.uint,  # ev_op_iterations
-        nb.types.uint,  # ev_op_max_order_qcd
-        nb.types.uint,  # sv_mode_num
-        nb.types.bool_,  # is_threshold
-    ),
-    cache=False,
-)
-def cb_quad_ker_qcd(
-    re_gamma_raw,
-    im_gamma_raw,
-    re_n,
-    im_n,
-    re_jac,
-    im_jac,
-    order_qcd,
-    is_singlet,
-    mode0,
-    mode1,
-    nf,
-    is_log,
-    logx,
-    areas_raw,
-    areas_x,
-    areas_y,
-    L,
-    _method_num,
-    as1,
-    as0,
-    ev_op_iterations,
-    ev_op_max_order_qcd,
-    _sv_mode_num,
-    is_threshold,
-):
-    """C Callback inside integration kernel."""
-    n = re_n + im_n * 1j
-    jac = re_jac + im_jac * 1j
-    areas = nb.carray(areas_raw, (areas_x, areas_y))
-    pj = interpolation.evaluate_grid(n, is_log, logx, areas)
-    method = "iterate-expanded"
-    sv_mode = sv.Modes.exponentiated
-    order = (order_qcd, 0)
-    ev_op_max_order = (ev_op_max_order_qcd, 0)
-    if is_singlet:
-        re_gamma_singlet = nb.carray(re_gamma_raw, (order_qcd, 2, 2))
-        im_gamma_singlet = nb.carray(im_gamma_raw, (order_qcd, 2, 2))
-        gamma_singlet = re_gamma_singlet + im_gamma_singlet * 1j
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_singlet = sv.exponentiated.gamma_variation(
-                gamma_singlet, order, nf, L
-            )
-        ker = s.dispatcher(
-            order,
-            method,
-            gamma_singlet,
-            as1,
-            as0,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv.expanded.singlet_variation(gamma_singlet, as1, order, nf, L, dim=2)
-            ) @ np.ascontiguousarray(ker)
-        ker = select_singlet_element(ker, mode0, mode1)
-    else:
-        re_gamma_ns = nb.carray(re_gamma_raw, order_qcd)
-        im_gamma_ns = nb.carray(im_gamma_raw, order_qcd)
-        gamma_ns = re_gamma_ns + im_gamma_ns * 1j
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_ns = sv.exponentiated.gamma_variation(gamma_ns, order, nf, L)
-        ker = ns.dispatcher(
-            order,
-            method,
-            gamma_ns,
-            as1,
-            as0,
-            nf,
-            ev_op_iterations,
-        )
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = sv.expanded.non_singlet_variation(gamma_ns, as1, order, nf, L) * ker
-    # recombine everything
-    res = ker * pj * jac
-    return np.real(res)
-
-
-@nb.njit()
-def quad_ker_qcd_ns(
-    order,
-    mode,
-    n,
-    nf,
-    quad_ker_qcd_ns_address,
-    L,
-    as1,
-    as0,
-    ev_op_iterations,
-    is_threshold,
-):
-    """Compute ker in C."""
-    # bypass complex and pointers
-    method_num = 1
-    sv_mode_num = 1
-    res = c_quad_ker_qcd_ns(
-        order[0],
-        mode,
-        np.real(n),
-        np.imag(n),
-        nf,
-        quad_ker_qcd_ns_address,
-        L,
-        method_num,
-        as1,
-        as0,
-        ev_op_iterations,
-        sv_mode_num,
-        is_threshold,
-        re_double_address,
-        im_double_address,
-    )
-    ker = read_complex(1)
-    return ker[0]
-
-
-@nb.njit(cache=True)
-def quad_ker_qed(
-    ker_base,
-    order,
-    mode0,
-    mode1,
-    method,
-    as_list,
-    mu2_from,
-    mu2_to,
-    a_half,
-    alphaem_running,
-    nf,
-    L,
-    ev_op_iterations,
-    ev_op_max_order,
-    sv_mode,
-    is_threshold,
-):
-    """Raw evolution kernel inside quad.
-
-    Parameters
-    ----------
-    ker_base : QuadKerBase
-        quad argument
-    order : int
-        perturbation order
-    mode0: int
-        pid for first sector element
-    mode1 : int
-        pid for second sector element
-    method : str
-        method
-    as1 : float
-        target coupling value
-    as0 : float
-        initial coupling value
-    mu2_from : float
-        initial value of mu2
-    mu2_from : float
-        final value of mu2
-    aem_list : list
-        list of electromagnetic coupling values
-    alphaem_running : bool
-        whether alphaem is running or not
-    nf : int
-        number of active flavors
-    L : float
-        logarithm of the squared ratio of factorization and renormalization scale
-    ev_op_iterations : int
-        number of evolution steps
-    ev_op_max_order : int
-        perturbative expansion order of U
-    sv_mode: int, `enum.IntEnum`
-        scale variation mode, see `eko.scale_variations.Modes`
-    is_threshold : boolean
-        is this an itermediate threshold operator?
-
-    Returns
-    -------
-    float
-        evaluated integration kernel
-    """
-    # compute the actual evolution kernel for QEDxQCD
-    if ker_base.is_QEDsinglet:
-        gamma_s = ad_us.gamma_singlet_qed(order, ker_base.n, nf)
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_s = sv.exponentiated.gamma_variation_qed(
-                gamma_s, order, nf, L, alphaem_running
-            )
-        ker = qed_s.dispatcher(
-            order,
-            method,
-            gamma_s,
-            as_list,
-            a_half,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        # TODO : in this way a_half[-1][1] is the aem value computed in
-        # the middle point of the last step. Instead we want aem computed in mu2_final.
-        # However the distance between the two is very small and affects only the running aem
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv.expanded.singlet_variation_qed(
-                    gamma_s, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
-                )
-            ) @ np.ascontiguousarray(ker)
-        ker = select_QEDsinglet_element(ker, mode0, mode1)
-    elif ker_base.is_QEDvalence:
-        gamma_v = ad_us.gamma_valence_qed(order, ker_base.n, nf)
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_v = sv.exponentiated.gamma_variation_qed(
-                gamma_v, order, nf, L, alphaem_running
-            )
-        ker = qed_v.dispatcher(
-            order,
-            method,
-            gamma_v,
-            as_list,
-            a_half,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv.expanded.valence_variation_qed(
-                    gamma_v, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
-                )
-            ) @ np.ascontiguousarray(ker)
-        ker = select_QEDvalence_element(ker, mode0, mode1)
-    else:
-        gamma_ns = ad_us.gamma_ns_qed(order, mode0, ker_base.n, nf)
-        # scale var exponentiated is directly applied on gamma
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_ns = sv.exponentiated.gamma_variation_qed(
-                gamma_ns, order, nf, L, alphaem_running
-            )
-        ker = qed_ns.dispatcher(
-            order,
-            method,
-            gamma_ns,
-            as_list,
-            a_half[:, 1],
-            alphaem_running,
-            nf,
-            ev_op_iterations,
-            mu2_from,
-            mu2_to,
-        )
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = (
-                sv.expanded.non_singlet_variation_qed(
-                    gamma_ns, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
-                )
-                * ker
-            )
-    return ker
 
 
 class Operator(sv.ModeMixin):
@@ -1120,7 +427,7 @@ class Operator(sv.ModeMixin):
                 cfg.mode0 = label[0]
                 cfg.mode1 = label[1]
                 user_data = ctypes.cast(ctypes.pointer(cfg), ctypes.c_void_p)
-                func = LowLevelCallable(c_quad_ker_qcd2, user_data)
+                func = LowLevelCallable(c_quad_ker_qcd, user_data)
                 res = integrate.quad(
                     func,
                     0.5,
