+++ src/ekore/anomalous_dimensions/unpolarized/space_like/__init__.py
@@ -20,49 +20,8 @@ import numpy as np
 
 from eko import constants
 
-from ....cffilib import (
-    c_ad_us_gamma_ns,
-    c_ad_us_gamma_singlet,
-    im_double_address,
-    re_double_address,
-    read_complex,
-)
 from ....harmonics import cache as c
 from . import aem1, aem2, as1, as1aem1, as2, as3, as4
-
-
-@nb.njit()
-def c_gamma_ns(order, mode, n, nf):
-    """Compute gamma_ns from C."""
-    # bypass complex and pointers
-    res = c_ad_us_gamma_ns(
-        order[0],
-        mode,
-        np.real(n),
-        np.imag(n),
-        nf,
-        re_double_address,
-        im_double_address,
-    )
-    gamma_ns = read_complex(order[0])
-    return gamma_ns
-
-
-@nb.njit()
-def c_gamma_singlet(order, n, nf, _n3lo_ad_variation):
-    """Compute gamma_singlet from C."""
-    # bypass complex and pointers
-    res = c_ad_us_gamma_singlet(
-        order[0],
-        np.real(n),
-        np.imag(n),
-        nf,
-        re_double_address,
-        im_double_address,
-    )
-    size = order[0] * 4
-    gamma_s = read_complex(size)
-    return gamma_s.reshape(order[0], 2, 2)
 
 
 @nb.njit(cache=True)
